Changed int variables to doubles.

+++ lowend/src/Main.java
@@ -8,9 +8,11 @@ public class Main {
         Scanner input = new Scanner(System.in);
 
         System.out.println("Welcome to Low-End Calculator!");
+
 while(true){
+
         System.out.println("Please enter a number.");
-        int num1 = input.nextInt();
+        double num1 = input.nextInt();
         System.out.println(num1);
 
         // prompt for operator
